Move Elementor files into the folder
Saves some searching around for Elementor specific JS code
