+++ packages/js/src/elementor/initialize.js
@@ -1,17 +1,17 @@
 import { dispatch } from "@wordpress/data";
 import { doAction } from "@wordpress/hooks";
-import initializeAiGenerator from "./ai-generator/initialize";
-import initEditorStore from "./elementor/initializers/editor-store";
-import initHighlightFocusKeyphraseForms from "./elementor/initializers/highlightFocusKeyphraseForms";
-import initializeIntroduction from "./elementor/initializers/introduction";
-import initializeIntroductionEditorV2 from "./elementor/initializers/introduction-editor-v2";
-import { applyModifications, pluginReady, pluginReloaded, registerModification, registerPlugin } from "./initializers/pluggable";
-import initializeUsedKeywords from "./elementor/initializers/used-keywords-assessment";
-import initReplaceVarPlugin, { addReplacement, ReplaceVar } from "./elementor/replaceVars/elementor-replacevar-plugin";
-import initAnalysis, { collectData } from "./initializers/analysis";
-import initElementorEditorIntegration from "./initializers/elementor-editor-integration";
-import initializeInsights from "./insights/initializer";
-import initElementorWatcher from "./watchers/elementorWatcher";
+import initializeAiGenerator from "../ai-generator/initialize";
+import initAnalysis, { collectData } from "../initializers/analysis";
+import { applyModifications, pluginReady, pluginReloaded, registerModification, registerPlugin } from "../initializers/pluggable";
+import initializeInsights from "../insights/initializer";
+import initElementorEditorIntegration from "./initializers/editor-integration";
+import initEditorStore from "./initializers/editor-store";
+import initElementorWatcher from "./initializers/editor-watcher";
+import initHighlightFocusKeyphraseForms from "./initializers/highlightFocusKeyphraseForms";
+import initializeIntroduction from "./initializers/introduction";
+import initializeIntroductionEditorV2 from "./initializers/introduction-editor-v2";
+import initializeUsedKeywords from "./initializers/used-keywords-assessment";
+import initReplaceVarPlugin, { addReplacement, ReplaceVar } from "./replaceVars/elementor-replacevar-plugin";
 
 /* eslint-disable complexity */
 /**
@@ -82,6 +82,7 @@ function initialize() {
 	// Offer an action after our load.
 	doAction( "yoast.elementor.loaded" );
 }
+
 /* eslint-enable complexity */
 
 // Wait on `window.elementor`.
